Comment functionality and hidden projects
