+++ systems/proto/app/Resources/views/general/messagebox.html.twig
@@ -1,5 +1,5 @@
-    <section class="narrow centered ph2 pv3">
-      <div class="message-box">
+    <section class="narrow centered ph2">
+      <div class="message-box mb3">
         <p><span class="fa fa-info-circle" aria-hidden="true"></span>{{ message }}</p>
       </div>
     </section>
